Created nuclear_module.py
This is my experimentation at learning about modules and how to import them into other programmes. It's going well so far :)

I've added the if __name__ == "__main__": line to binding_energy.py and nuclear_radii.py to prevent them running when called as a module. I will look into using this methodology to create a programme that returns all the possible information about a nuclei based on some basic input.